hci: allow for both ninafw and pure hci uart adapter implementations
